refactor client-side websocket into its own hook
